backend/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
crewai==1.8.1
langchain-openai==0.1.25
python-dotenv==1.1.1